third_party/blink/web_tests/wpt_internal/route/parsing/active-navigation-valid.html

<!DOCTYPE html>
<title>Serialization of @navigation</title>
<link rel="author" title="Morten Stenshorne" href="mailto:mstensho@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-navigation-1/#active-navigation-pseudo-class">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
  test(()=> {
    const canonical = ":active-navigation(with link-href) { }";
    test_valid_rule(canonical);
    test_valid_rule(":Active-navigation(){} ", canonical);
    test_valid_rule(":Active-navigation(Link-Href ) { }", canonical);
  }, "no arguments, link-href");

  test(()=> {
    const canonical = ":active-navigation(at link-href) { }";
    test_valid_rule(canonical);
    test_valid_rule(":Active-navigation( At ){} ", canonical);
    test_valid_rule(":Active-navigation( At  Link-Href ) { }", canonical);
  }, "explicit relation");

  test(()=> {
    const canonical = ":active-navigation(at --rte) { }";
    test_valid_rule(canonical);
    test_valid_rule(":Active-navigation( At --rte ){} ", canonical);
  }, "explicit location");
</script>
